Clean up endpoints with broken imports
