+++ src/components/UIUC-Components/DocumentsCard.tsx
@@ -4,7 +4,7 @@ import { IconFileExport } from '@tabler/icons-react'
 import { montserrat_heading, montserrat_paragraph } from 'fonts'
 import { useRouter } from 'next/router'
 import { useState } from 'react'
-import { handleExport } from '~/pages/api/UIUC-api/exportAllDocuments'
+import { handleExport } from '~/pages/util/handleExport'
 import { type CourseMetadata } from '~/types/courseMetadata'
 import { useResponsiveCardWidth } from '~/utils/responsiveGrid'
 import { showToastOnUpdate } from './MakeQueryAnalysisPage'
